got rid of repetitive code when rendering board squares

+++ src/ui/rendering/pieces.rs
@@ -85,11 +85,11 @@ impl Textures {
     }
 }
 
-pub fn draw_piece(textures: &Textures, x: i32, y: i32, piece: Piece) {
+pub fn draw_piece(textures: &Textures, x: usize, y: usize, piece: Piece) {
     draw_texture_ex(
         textures.get_texture(&piece),
-        (x * constants::board::SQUARE_SIZE + constants::board::PIECE_PADDING) as f32,
-        (y * constants::board::SQUARE_SIZE + constants::board::PIECE_PADDING) as f32,
+        (x as i32 * constants::board::SQUARE_SIZE + constants::board::PIECE_PADDING) as f32,
+        (y as i32 * constants::board::SQUARE_SIZE + constants::board::PIECE_PADDING) as f32,
         WHITE,
         DrawTextureParams {
             dest_size: Some(Vec2::from((
